AtmosphereScattering: Moved some classes.
