+++ source/Plugins/Generic/AtmosphereScattering/AtmosphereBackgroundPass.cpp
@@ -140,7 +140,7 @@ namespace atmosphere_scattering
 					auto clipSpace = writer.declLocale( "clipSpace"
 						, vec3( ( pixPos / targetSize ) * vec2( 2.0_f, -2.0_f ) - vec2( 1.0_f, -1.0_f ), 1.0_f ) );
 					auto hPos = writer.declLocale( "hPos"
-						, c3d_cameraData.invViewProj * vec4( clipSpace, 1.0_f ) );
+						, c3d_cameraData.projToWorld( vec4( clipSpace, 1.0_f ) ) );
 
 					auto worldDir = writer.declLocale( "worldDir"
 						, normalize( hPos.xyz() / hPos.w() - c3d_cameraData.position ) );
@@ -202,7 +202,7 @@ namespace atmosphere_scattering
 							&& "The FASTAERIALPERSPECTIVE_ENABLED path does not support COLORED_TRANSMITTANCE_ENABLED." );
 						clipSpace = vec3( ( pixPos / targetSize ) * vec2( 2.0_f, -2.0_f ) - vec2( 1.0_f, -1.0_f ), depthBufferValue );
 						auto depthBufferWorldPos = writer.declLocale( "depthBufferWorldPos"
-							, c3d_cameraData.invViewProj * vec4( clipSpace, 1.0_f ) );
+							, c3d_cameraData.projToWorld( vec4( clipSpace, 1.0_f ) ) );
 						depthBufferWorldPos /= depthBufferWorldPos.w();
 						auto tDepth = writer.declLocale( "tDepth"
 							, length( depthBufferWorldPos.xyz() - ( worldPos + vec3( 0.0_f, 0.0_f, -c3d_atmosphereData.bottomRadius ) ) ) );
